vm1/consumer/requirements.txt
kafka-python
six
python-dotenv
fastapi
uvicorn[standard]
websockets